src/test/java/com/crud/CrudApp/UserServiceTests.java
package com.crud.CrudApp;

import com.crud.CrudApp.User.User;
import com.crud.CrudApp.User.UserService;
import org.junit.jupiter.api.Disabled;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.ArgumentsSource;
import org.junit.jupiter.params.provider.ValueSource;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;

import static org.junit.jupiter.api.Assertions.*;

@SpringBootTest
public class UserServiceTests {
    @Autowired
    private UserService userService;

    @Disabled
    @ParameterizedTest
    @ArgumentsSource(UserArguementsProvider.class)
    public void testSaveNewUser(User user) {
        assertTrue(userService.saveNewUser(user));
    }

    @ParameterizedTest
    @ValueSource(strings = {
            "Raman"
    })
    public void testGetByUsername(String username) {
        assertNotNull(userService.getByUsername(username));
    }
}
